simple/universal example same as other ones; move its previous contents under general examples into z_custom_serializer.cxx;
