+++ examples/universal/z_sub_attachment.cxx
@@ -72,8 +72,7 @@ int _main(int argc, char **argv) {
 
     std::cout << "Declaring Subscriber on '" << keyexpr.as_string_view() << "'..." << std::endl;
     auto subscriber = session.declare_subscriber(keyexpr, data_handler, closures::none);
-
-    std::cout << "Subscriber on '" << keyexpr.as_string_view() << "' declared" << std::endl;
+    std::cout << "Subscriber on '" << subscriber.get_keyexpr().as_string_view() << "' declared" << std::endl;
 
     std::cout << "Press CTRL-C to quit...\n";
     while (true) {
